% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.ss}
\alias{dat.ss}
\title{dat.ss}
\format{
single-cell mRNA table.
}
\usage{
dat.ss
}
\description{
The single-cell mRNA dataset of 96 samples (rows), and 100 genes (columns).
}
\examples{
dat.ss
dim(dat.ss)


}
\keyword{datasets}
