% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{dirichletVariation}
\alias{dirichletVariation}
\title{Dirichlet Variation}
\usage{
dirichletVariation(counts)
}
\arguments{
\item{counts}{Dataset of compositional data}
}
\value{
\code{NULL}
}
\description{
Estimates the variation matrix for empirical compositional data which is based
on the Dirichlet distribution.
}
