% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{grad}
\alias{grad}
\title{Gradient of Normal Random Variables}
\usage{
grad(v, y, ni, mu, Sigma.inv)
}
\arguments{
\item{v}{Additive-log ratio (alr) transform of y}

\item{y}{Compositional data set}

\item{ni}{Row sums of y?}

\item{mu}{Mu vector of y}

\item{Sigma.inv}{Sigma inverse matrix of y}
}
\value{
The gradient vector of Normal random variables
}
\description{
Calculates the gradient of the normal random variables, on the logit scale.
}
