% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{hess}
\alias{hess}
\title{Hessian}
\usage{
hess(v, ni, Sigma.inv)
}
\arguments{
\item{v}{Compositional dataset which has been transformed by the additive log ratio}

\item{ni}{Row sums of the raw data ?}

\item{Sigma.inv}{Inverse of the Sigma matrix}
}
\value{
The hessian matrix.
}
\description{
Calculates the hessian matrix.
}
