% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{logitNormalVariation}
\alias{logitNormalVariation}
\title{Logit Normal Variation}
\usage{
logitNormalVariation(
  mu,
  Sigma,
  type = c("standard", "phi", "phis", "rho"),
  order = c("first", "second")
)
}
\arguments{
\item{mu}{The mle estimate of the mu matrix}

\item{Sigma}{The mle estimate of the Sigma matrix}

\item{type}{Type of variation metric to be calculated: \code{standard}, \code{phi},
\code{phis} (a symmetrical version of \code{phi}), or \code{rho}}

\item{order}{The order of the Taylor-series approximation to be used in the 
estimation}
}
\value{
An estimation of the variation matrix, \code{V}.
}
\description{
Estimates the variation matrix of count-compositional data 
based on a multinomial logit-Normal distribution. Estimate is performed using 
only the parameters of the distribution.
}
\examples{
mu.hat <- mle$est.min$mu
Sigma.hat <- mle$est.min$Sigma

logitNormalVariation(mu.hat, Sigma.hat) #Standard logit-Normal estimates of variance
logitNormalVariation(mu.hat, Sigma.hat, type="phi", order="second") #Logit-Normal based estimate of phi
logitNormalVariation(mu.hat, Sigma.hat, type="phis", order="second") #Logit-Normal based estimate of phis
logitNormalVariation(mu.hat, Sigma.hat, type="rho", order="second") #Logit-Normal based estimate of rho

}
