% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{logitNormalVariationOLD}
\alias{logitNormalVariationOLD}
\title{Logit Normal Variation (Old Version)}
\usage{
logitNormalVariationOLD(
  mu,
  Sigma,
  lmu,
  lsigma,
  type = c("standard", "phi", "rho")
)
}
\arguments{
\item{mu}{?}

\item{Sigma}{?}

\item{lmu}{?}

\item{lsigma}{?}

\item{type}{Type of variation metric to be calculated: \code{standard}, \code{phi},
or \code{rho}}
}
\value{
An estimation of the variation matrix, \code{V}.
}
\description{
Estimates the empirical variation matrix of count data that follows a multinomial 
logit-Normal distribution. Differs from the function \code{logitNormalVariation()}
in that this function does not use a Taylor-series approximation.
}
