% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{logLik}
\alias{logLik}
\title{Log-Likelihood}
\usage{
logLik(v, y, ni, S, invSigma)
}
\arguments{
\item{v}{The additive logratio transform of y}

\item{y}{Compositional dataset}

\item{ni}{The row sums of y}

\item{S}{Covariance of \code{v}}

\item{invSigma}{The inverse of the Sigma matrix}
}
\value{
The estimated log-likelihood under the Multinomial logit-Normal distribution.
}
\description{
Calculates the log-likelihood, under the multinomial logit-Normal model.
}
