% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{logLikG}
\alias{logLikG}
\title{Gaussian Log-Likelihood}
\usage{
logLikG(v, S, invSigma)
}
\arguments{
\item{v}{Count-compositional dataset which has been transformed by the additive
logratio}

\item{S}{Covariance of \code{v}}

\item{invSigma}{Inverse of the Sigma matrix}
}
\value{
The estimated Gaussian log-likelihood under the Multinomial logit-Normal distribution.
}
\description{
Calculates the Gaussian log-likelihood, under the multinomial logit-Normal model.
}
