% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{logVarMC}
\alias{logVarMC}
\title{Monte Carlo Logx Variance-Covariance}
\usage{
logVarMC(mu, Sigma, K = 1e+05)
}
\arguments{
\item{mu}{The mean vector of the underlying distribution}

\item{Sigma}{The variance matrix of the underlying distribution}

\item{K}{Number of samples}
}
\value{
The estimated variance-covariance matrix, \code{logx}
}
\description{
Function which estimates the variance0covariance of the log of the data, using Monte
Carlo integration.
}
