% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{logVarUnscented}
\alias{logVarUnscented}
\title{Unscented Logx Variance-Covariance}
\usage{
logVarUnscented(
  mu,
  Sigma,
  transf = c("alr", "clr"),
  alpha = 0.001,
  beta = 2,
  kappa = 0
)
}
\arguments{
\item{mu}{The mean vector of the underlying distribution}

\item{Sigma}{The sigma matrix of the underlying distribution}

\item{transf}{The desired transformation. If \code{transf="alr"} the inverse 
additive log-ratio transformation is applied. If \code{transf="clr"} the
inverse centered log-ratio transformation is applied.}

\item{alpha}{Parameter which controls the spread of the Sigma points}

\item{beta}{Parameter which compensates for the distribution}

\item{kappa}{Scaling Parameter}
}
\value{
The estimated variance-covariance matrix, logx.
}
\description{
Function which estimates the variance-covariance of the log of the data, using an 
unscented transformation.
}
\references{
Extended and Unscented Kalman Filter Algorithms for Online State Estimation. 
(2022). MathWorks. https://www.bibliography.com/apa/apa-reference-page-examples-and-format-guide/

Hendeby, G., & Gustafsson, F. On Nonlinear Transformations Of Gaussian Distributions. 
https://users.isy.liu.se/en/rt/fredrik/reports/07SSPut.pdf
}
