% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{logVarUnscentedW}
\alias{logVarUnscentedW}
\title{Unscented Logx Variance-Covariance   ???What is W for?}
\usage{
logVarUnscentedW(mu, Sigma, transf = c("alr", "clr"))
}
\arguments{
\item{mu}{The mean vector of the underlying distribution}

\item{Sigma}{The sigma matrix of the underlying distribution}

\item{transf}{The desired transformation. If \code(transf="alr") the inverse 
additive logratio transformation is applied. If \code(transf="clr") the
inverse centered logratio transformation is applied.}
}
\value{
The estimated variance-covariance matrix, logx.
}
\description{
Function which estimates the variance-covariance of the log of the data, using an 
unscented transformation.
}
