% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{MCVariation}
\alias{MCVariation}
\title{Monte Carlo Variation}
\usage{
MCVariation(
  mu = NULL,
  Sigma = NULL,
  x = NULL,
  K = 1e+06,
  type = c("standard", "phi", "phis", "rho", "logx")
)
}
\arguments{
\item{mu}{The mean matrix of the underlying distribution}

\item{Sigma}{The variance matrix of the underlying distribution}

\item{x}{A sample from the multinomial logit-Normal model. If \code{mu=NULL} 
and \code{Sigma=NULL}, then x must be provided}

\item{K}{Number of Monte Carlo samples to generate}

\item{type}{Type of variation metric to be calculated: \code{standard}, \code{phi},
\code{phis} (a symmetrical version of \code{phi}), \code{rho}, or \code{logx}}
}
\value{
The variance matrix, \code{v}.
}
\description{
Calculates the variation matrix using Monte Carlo integration.
}
