% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mle}
\alias{mle}
\title{MLE}
\format{
Gives mle estimates of \code{mu} and \code{Sigma} of \code{dat.ss}.
}
\usage{
mle
}
\description{
Maximum likelihood estimates from dat.ss, such as the \code{mean} and 
\code{Sigma}. MLE based on the multinomial logit-Normal model.
}
\examples{
mu <- mle$est.min$mu
Sigma <- mle$est.min$Sigma


}
\keyword{datasets}
