% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{mlePath}
\alias{mlePath}
\title{Maximum Likelihood Estimator Paths}
\usage{
mlePath(
  y,
  max.iter = 10000,
  max.iter.nr = 100,
  tol = 1e-06,
  tol.nr = 1e-06,
  lambda.gl = NULL,
  lambda.min.ratio = 0.1,
  n.lambda = 1,
  n.cores = NULL,
  gamma = 0.1
)
}
\arguments{
\item{y}{Compositional dataset}

\item{max.iter}{Maximum number of iterations}

\item{max.iter.nr}{Maximum number of Newton-Raphson iterations}

\item{tol}{Stopping rule}

\item{tol.nr}{Stopping rule for the Newton Raphson algorithm}

\item{lambda.gl}{Vector of penalization parameters lambda, for graphical lasso}

\item{lambda.min.ratio}{Minimum lambda ratio of the maximum lambda, 
used for the sequence of lambdas}

\item{n.lambda}{Number of lambda to evaluate differnet paths for}

\item{n.cores}{Number of cores to use for parallel computation}

\item{gamma}{Gamma value for EBIC calculation of the log-likelihood}
}
\value{
The MLE estimates of y for each element lambda of lambda.gl, (\code{est}); 
the value of the estimates which produce the minimum EBIC, (\code{est.min}); 
the vector of lambdas used for graphical lasso, (\code{lambda.gl}); the index of 
the minimum EBIC (extended Bayesian information criterion), (\code{min.idx}); 
vector containg the EBIC for each lambda, (\code{ebic}).
}
\description{
Calculates the maximum likelihood estimates of the parameters for various 'paths'
of the penalization parameter \code{lambda}.
}
