% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{pluginVariation}
\alias{pluginVariation}
\title{Plugin Variation}
\usage{
pluginVariation(counts)
}
\arguments{
\item{counts}{Dataset of compositional data}
}
\value{
\code{NULL}
}
\description{
Estimates the variation matrix for empirical compositional data, ?plugin.
}
