% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(x, y)
}
\arguments{
\item{x}{Value one for comparison (a scaler, vector, or matrix)}

\item{y}{Value two for comparison (a scaler, vector, or matrix). Must be the 
same form as \code{x}}
}
\value{
A single scaler value, the RMSE.
}
\description{
Calculates the root mean square error (RMSE) between two values (i.e. scalers, 
vectors, or matrices).
}
\examples{
x <- sample(1:20, size = 4)
y <- sample(1:20, size = 4)

rmse(x,y)

}
