% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{rmse_by_row}
\alias{rmse_by_row}
\title{Row Root Mean Square Error}
\usage{
rmse_by_row(x, y)
}
\arguments{
\item{x}{First matrix for comparison}

\item{y}{Second matrix for comparison. Must be the same dimensions as \code{x}}
}
\value{
A vector. Element \code{i} of the resultant vector is the RMSE which 
compares row \code{i} of matrix \code{x}, and row \code{i} of matrix \code{y}.
}
\description{
Calculates the root mean square error (RMSE) between each row of two matrices.
}
\examples{
x <- matrix(sample(1:500, size = 21), nrow=3)
y <- matrix(sample(1:500, size = 21), nrow=3)

rmse(x,y)

}
