% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{wrapMLE}
\alias{wrapMLE}
\title{Wrapper for MleLR()}
\usage{
wrapMLE(x)
}
\arguments{
\item{x}{Input argument for the \code{MLE Function}}
}
\description{
Executes a function call to \code{mlePath()}. Helps with the \code{mclapply()} 
within \code{mlePath()}
}
