% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{ebic}
\alias{ebic}
\title{Extended Bayesian Information Criterion}
\usage{
ebic(l, n, d, df, gamma)
}
\arguments{
\item{l}{Log-likelihood estimates of the model}

\item{n}{Number of rows of the data set for which the log-likelihood has been 
calculated}

\item{d}{?}

\item{df}{Degrees of freedom}

\item{gamma}{?}
}
\value{
The value of the EBIC.
}
\description{
Calculates the Extende
d Bayesian Information Criterion (EBIC) of a model.
Used for model selection.
}
