% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{g}
\alias{g}
\title{Log of the Inverse Additive Log-ratio}
\usage{
g(x)
}
\arguments{
\item{x}{Compositional data vector which has already been transformed by the 
additive log-ratio}
}
\value{
A vector which is the log of the inverse of the data which has been
transformed by the additive logratio transformation.
}
\description{
Function which takes the logarithm of a vector, after the inverse additive 
log-ratio transformation has been applied to it.
}
\examples{
vec <- sample(1:250, 9)
ref <- vec[length(vec)]
alr.vec <- log(vec[-length(vec)]/ref)

g(alr.vec)

}
