% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{logitNormalVariation}
\alias{logitNormalVariation}
\title{Logit Normal Variation}
\usage{
logitNormalVariation(
  mu,
  Sigma,
  type = c("standard", "phi", "phis", "rho"),
  order = c("first", "second")
)
}
\arguments{
\item{mu}{The mle estimate of the mu matrix}

\item{Sigma}{The mle estimate of the sigma matrix}

\item{type}{Type of variation metric to be calculated: \code{standard}, \code{phi},
\code{phis} (a symmetrical version of \code{phi}), or \code{rho}}

\item{order}{The order of the Taylor-series approximation to be used in the 
estimation}
}
\value{
An estimation of the variation matrix, \code{V}.
}
\description{
Estimates the empirical variation matrix of count data that follows a multinomial
logit-Normal distribution.
}
