% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{logLik}
\alias{logLik}
\title{Log-Likelihood}
\usage{
logLik(v, y, ni, S, invSigma)
}
\arguments{
\item{v}{The additive log ratio transform of y}

\item{y}{Compositional data set}

\item{ni}{The row sums of y?}

\item{S}{?}

\item{invSigma}{The inverse of the Sigma matrix}
}
\value{
The estimated log-likelihood under the Multinomial Logit-Normal distribution.
}
\description{
Calculates the log-likelihood.
}
