% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{logVarTaylorFull}
\alias{logVarTaylorFull}
\title{Full Logx Variance-Covariance}
\usage{
logVarTaylorFull(
  mu,
  Sigma,
  transf = c("alr", "clr"),
  order = c("first", "second")
)
}
\arguments{
\item{mu}{The mean vector of the underlying distribution}

\item{Sigma}{The sigma matrix of the underlying distribution}

\item{transf}{The desired transformation. If \code{transf="alr"} the inverse 
additive logratio transformation is applied. If \code{transf="clr"} the
inverse centered logratio transformation is applied.}

\item{order}{The desired order of the Taylor Series approximation}
}
\value{
The estimated variance-covariance matrix, \code{logx}.
}
\description{
Function which estimates the variance-covariance of the log of the data, using a 
Taylor-series approximation. This function differs from \code{Logx Variance} in
that the resultant matrix includes a reference category.
}
