% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{MCSample}
\alias{MCSample}
\title{Monte Carlo Sample}
\usage{
MCSample(mu, Sigma, K = 1)
}
\arguments{
\item{mu}{Mean matrix of the underlying distribution}

\item{Sigma}{Variance matrix of the underlying distribution}

\item{K}{Number of samples to generate}
}
\value{
K samples from the multinomial logit-normal model. The number of features
in the sample is of length(\code{mu})+1.
}
\description{
Generates a Monte Carlo sample based on the multinomial logit-normal model.
}
