% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{MCVariation}
\alias{MCVariation}
\title{Monte Carlo Variation}
\usage{
MCVariation(
  mu = NULL,
  Sigma = NULL,
  x = NULL,
  K = 1e+06,
  type = c("standard", "phi", "phis", "rho", "logx")
)
}
\arguments{
\item{mu}{The mean vector of a dataset following the multinomial logit-Normal model}

\item{Sigma}{The variance matrix of a dataset following the multinomial logit-Normal model}

\item{x}{A sample from the multinomial logit-Normal model. If \code{mu=NULL} 
and \code{Sigma=NULL}, then x must be provided}

\item{K}{Number of Monte Carlo samples to generate}

\item{type}{Type of variation metric to be calculated: \code{standard}, \code{phi},
\code{phis} (a symmetrical version of \code{phi}), \code{rho}, or \code{logx}}
}
\value{
The variance matrix, \code{v}.
}
\description{
Estimates the "true" values of the variation matrix for count-compositional 
data based on the multinomial logit-Normal model, using large-sample Monte-Carlo methods.
}
\examples{
mu <- mle$est.min$mu
Sigma <- mle$est.min$Sigma

MCVariation(mu, Sigma, type="standard") #Monte Carlo estimate of the standard variation, using mu and Sigma    
MCVariation(x=dat.ss, type="standard")   #Monte Carlo estimate of the standard variation, using x


}
