% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_multinom_logit.R
\name{mleLR}
\alias{mleLR}
\title{Maximum Likelihood Estimate}
\usage{
mleLR(
  y,
  max.iter = 10000,
  max.iter.nr = 100,
  tol = 1e-06,
  tol.nr = 1e-06,
  lambda.gl = 0,
  gamma = 0.1
)
}
\arguments{
\item{y}{Compositional dataset}

\item{max.iter}{Maximum number of iterations}

\item{max.iter.nr}{Maximum number of Newton-Raphson iterations}

\item{tol}{Stopping rule}

\item{tol.nr}{Stopping rule for the Newton-Raphson algorithm}

\item{lambda.gl}{Penalization parameter lambda, for the graphical lasso}

\item{gamma}{Gamma value for EBIC calculation of the log-likelihood}
}
\value{

}
\description{
Function which returns maximum likelihood estimates of parameters given a compositional
dataset. The MLE procedure is based on the Multinomial Logit-Normal distribution, 
using the EM algorithm from Hoff (2003).
}
