% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{naiveVariation}
\alias{naiveVariation}
\title{Naive Variation}
\usage{
naiveVariation(
  counts,
  pseudo.count = 0,
  type = c("standard", "phi", "phis", "rho", "logx"),
  use = "everything",
  set.inf.na = TRUE,
  already.log = FALSE
)
}
\arguments{
\item{counts}{Count-compositional dataset}

\item{pseudo.count}{Scaler value added to the data matrix to prevent infinite 
values caused by taking the log of the counts}

\item{type}{Type of variation metric to be calculated: \code{standard}, \code{phi},
\code{phis} (a symmetrical version of \code{phi}), \code{rho}, or \code{logx}}

\item{use}{If equal to \code{"everything"} and there are no infinite values after
taking the log the calculation will use all data. If equal to \code{"everything"} 
and there are infinite values after taking the log it is recommended to  run 
the function again, instead setting the \code{use} parameter to \code{"pairwise.complete.obs"}}

\item{set.inf.na}{If \code{TRUE}, sets any infinite values in \code{counts} to \code{NA}}

\item{already.log}{If \code{FALSE}, the counts have not been transformed by 
by the log. This transformation is of the form \eqn{log(frac{X_{ij}}{s_{i}})}, where 
\eqn{s_{i} = \sum{n=1}^{j} X_{in}}, where \eqn{X_{ij}} is element \eqn{counts[i,j]}}
}
\value{
The naive variation matrix, \code{v}.
}
\description{
Estimates the variation matrix of count compositional data using the proportions
of a count compositional dataset.
}
\examples{
n.g <- ncol(dat.ss)

naiveVariation(dat.ss)[-n.g,-n.g] #Standard naive estimate of the variance
naiveVariation(dat.ss, type="phi")[-n.g,-n.g] #Logit-Normal based naive estimate of phi
naiveVariation(dat.ss, type="phis")[-n.g,-n.g] #Logit-Normal based naive estimate of phis
naiveVariation(dat.ss, type="rho")[-n.g,-n.g] #Logit-Normal based naive estimate of rho

}
