% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_functions.R
\name{varEst}
\alias{varEst}
\title{Estimated Variation Matrix}
\usage{
varEst(
  counts,
  p.model = c("logitNormal", "dirichlet", "plugin"),
  type = c("standard", "phi", "rho"),
  refColumn = NULL
)
}
\arguments{
\item{counts}{Dataset of compositional data}

\item{p.model}{Probability model for the counts: (\code{logitNormal}, \code{dirichlet},
or \code{plugin})}

\item{type}{Type of variation metric to be calculated: \code{standard}, 
\code{phi}, or \code{rho}}

\item{refColumn}{The reference column to be used for the counts matrix}
}
\value{
The estimated variation matrix for the counts. May be a \code{standard},
\code{phi}, or \code{rho} variation matrix depending on \code{type} specified.
}
\description{
Estimates the variation matrix of compositional count data.
}
